pyproject.toml
[tool.poetry]
name = "kafka-module"
version = "0.1.3"
description = ""
authors = ["parkyoungjin-888 <pyj8813@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kafka-python-ng = "2.2.2"
data-model-module = { git = "https://github.com/parkyoungjin-888/data_model_module.git" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
